+++ types.ts
@@ -77,6 +77,11 @@ export const POST_PACKET_SCHEMA = z.object({
   val: POST_SCHEMA,
 });
 
+export const UPDATE_POST_PACKET_SCHEMA = z.object({
+  cmd: z.literal("update_post"),
+  val: POST_SCHEMA
+})
+
 /** An attachment as returned from the uploading API. */
 export type UploadsAttachment = {
   bucket: string;